bot: improve diff check
new_text from data.lua might contain trailing newline character.  This
shows in the diff, but is not actually recorded by MediaWiki.  So
rstrip() the new_text so that this diff hunk is not printed.

We have to drop parameter "context" from showDiff invocation, because it
is deprecated, and it causes IndexError when new_text is rstrip()ed.

Move diff above the summary, while we're here, to mimic MediaWiki web
interface UI of showing preview of changes.