run.sh: log in on every invocation
It seems that Fandom changed how cookies or whatever required for
performing edits work.  It used to be that I only needed to call the
"login" script once.  The bot has been erroring out for the last seven
comics.

Call the "login" script of Pywikibot on every invocation of the "run.sh"
script.